Implented Basic Equipables and Player Movement
System added for player to be able to have only one tool drain diesel at a time. Player Movement Revision is roughly 80% complete.

+++ Assets/Assets/Scripts/Damage Scripts/Damage Interfaces/Damager and Damageable Interfaces.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    public void TakeDamage(int damage_taken);
+    public void TakeDamage(float damage_taken);
 }
 
 public interface IDamager
