+++ Assets/Assets/Scripts/Enemy/GenericEnemy.cs
@@ -6,9 +6,9 @@ using UnityEngine;
 public class GenericEnemy : MonoBehaviour
 {
 
-    [SerializeField] private int health;
+    [SerializeField] private float health;
 
-    public void TakeMyDamage(int damage_taken)
+    public void TakeMyDamage(float damage_taken)
     {
         health -= damage_taken;
         Debug.Log("In Generic Enemy Script of Object" + this.gameObject.name + " " + health);
